Added Language updater script
Once for private use, now no more!

This script checks all the mod language files and updates
the mappings to the ones specified in the en_US.lang file.

Every deleted entry is also deleted on the other files, while comments
and blank lines are preserved. New translations are added and suffixed with
a ## NEEDS TRANSLATION ## entry, so that the language file updating can be
performed with a simple search.

This script has been tested and works as expected. Only problem is
the addition of the ## NEEDS TRANSLATION ## keyword at the end of the line
makes it appear in-game. I thought that this is not a huge problem, since
all translations should be updated before a release, but if this is
considered annoying, a simple update of the script (one character) will fix
this.

Warning! This script doesn't check the mod sources, so any
non-translated entry which is not present in en_US.lang (aka it shows with
its internal name in-game) will be ignored.

Signed-off by TheSilkMiner