Remove IGW Mod hard dependency
Instead of letting FML show the 'Missing mod screen', OpenMods-IGW
will now load and show a GUI which will let the user choose to
install IGW-Mod or simply continue without the benefits of this
mod.

Since hotload is disabled with mods, should the user choose to
install IGW, MC will open up IGW-Mod's page on Curse and then
close.

Last but not least, the dependency on IGW-Mod is kept as a light
one, which means OpenMods-IGW will load after it if present.

Signed-off by TheSilkMiner