Fix entity not showing when clicking on page link
When clicking on a page links which points to an entity, the
wiki should display the specified entity and show an
information page about it.

Before this commit, this was not the behaviour because the
research was only performed between registered ItemStacks,
rendering the entity if and only if it was the last thing
hovered before opening IGW's gui.

With this commit, this behavior is now fixed.

* Moved registration on a secondary method (see IPageInit.java)

  This means it will be easier to mantain since we will
  avoid code duplication.

* Rewrite ClientProxy detection of IGW-Mod

Signed-off by TheSilkMiner