Add hover text to warn user of web page opening
I thought it was a bit too rude to simply close Minecraft
and open a web page without warning the user. The text on
the button was too short: just a simple '(Opens browser)'
which did not answer some questions IMHO.

With the new hovering text, the warning will show up and
the URL of the page the button will open always appears,
regardless of translation, so that the user is always
warned and can check if the website we open is not a
malicious one (the two addresses correspond).

Signed-off by TheSilkMiner