Move common code to super class
Since OpenBlocks's wiki page contained a lot of code
that is useful to other classes, I decided to move
all the implementation to a super class, which can
then be used as a blueprint for all the other wiki
pages.

An example of the usefulness of this change is the
common Wiki page that will be loaded when the
'mono-tab' config option is enabled.

The implementation of the latter still needs some
work, though.

Also the restriction to the registration method in
ClientProxy hasn't been performed due to inheritance:
an OpenModsWikiPage is also an IWikiTab and I think
that limiting the method to the former is a bit too
much restrictive

Signed-off by TheSilkMiner