Full implementation of unique wiki tab
Now the implementation relies on IPositionProviders,
one for static pages and one for items. This allows a
finer control over the placement of the objects on the
wiki page.

Even if this change has been made to allow the
OpenModsCommonTab to function properly, it should be
considered still a WIP feature, even if available in
the configs.

Signed-off by TheSilkMiner