Add possibility for mod developers to specify custom versions
Some mods may not change a lot between two versions (such as
between 1.6 and 1.6.1), so having to update the wiki everytime
to reflect the changes and/or having to click through a GUI
everytime at launch. Using this custom annotation, you can
tell OpenMods-IGW to consider the installed mod version the
same as the version it was developed with. This would avoid
the GUI display and/or the various errors messages.

Using this method, only the change log will be out-of-date
(unless the developer forgot to edit the annotation's value,
but that's another story), but that is such a minor thing
that it can be safely ignored.

Signed-off by TheSilkMiner <sir.karati@outlook.com>