Add system detector service
This will allow us to activate special features when the environment where
the mod is run respects certain characteristics.

This does not mean this mod will provide special features for donators,
because that's against OpenMods policy. Instead the mod will activate
special debug features when run on the developers' systems.

The various features still have to be added.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>