Switch to static method invocation for mismatching mod entry
This will allow more future-proof code, allowing us to maintain backwards
compatibility with applications that use a previous API/prefab version
without resorting to various hacks involving ASM or reflection.

Also, this brings the construction of mismatching mod entries in line with
the other various construction ways used throurought this and other APIs,
such as Guava. E.g. a mod entry uses its "of" method to construct and cache
an entry; the construction of a Guava's Optional uses the static method "of"
and so on.

A side effect of this commit is also the caching ability of mismatching mod
entries.

-- INTERNAL NOTE --
Still no GPG sign, will be fixed as soon as possible

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>