Remove service caching from proxies
Services should never be cached because the underlying implementation could change
at any time, locking the mod in an invalid state. Given that, caching is not
suggested (or even allowed) for services, at least until service registering and
unregistering does not throw events which can be subscribed to (on TODO, LP).

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>