Add utility methods to system identifer service
The newly added methods allow for easier usage of the system service.
In fact, they automatically consider the current system as the one
the service should refer to, without having to populate manually a
SystemDetail everytime it is needed.

With this change, it is easier to write clearer and cleaner code.

All previous usages will still work correctly: no change has been
made to the original methods, which are still called by these new
methods: expects the same results and same exceptions, in case
something is wrong.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>