Allow users to enable various experimental features
Some of the various features developed may need more testing
other than the one a developer can offer and/or may still be
buggy and cause undefined behaviour.

This way, users who want the "bleeding edge" can still use
the latest features, while users who prefer to be on the
"safe edge" can still enjoy the mod.

Examples of experimental features may be the changelog in the
wiki pages (still a draft somewhere in my notes) or animated
GUI backgrounds (see TODO.MD for further details).

This option depends on the system identifier service: if the
service differs from the normal implementation (e.g. because
an addon substituted it), functionality is not always
guaranteed.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>