Abstract out last openmods.Log calls
This change adds for the first time a service which already
has an implementation. This new service allows an end user to
log anytime, without having to worry that the actual logging
service still has to be loaded or not.

The output is very similar to the one of openmods.Log, but
different in a variety of reasons to actually identify it as
a fallback. Otherwise, there would be no point is switching,
would it? :+1:

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>