Use separate image namespace
At the moment, this doesn't change anything.
If we will integrate with other mods, though, not having a namespace
for all the different mods may lead to duplicate names and thus we will
have no place to put the various images.
Separating the namespaces, i.e. placing all blocks from a mod within a
directory with that mod ID, will allow us to avoid that problem in the
future.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>