Porting step 3: Prettify code
I honestly thought that the previous step would have left behind a much
worse code base than the one we have here. In fact, everything is pretty
much the way it was before the porting, apart from a few "deprecated"
here and there that were fixed before and the entire registry stuff.

Ironically. I think that the registry by itself lead to better and
prettier code than the one that already was, so... yeah. A very small
step, but that leads us to a better modding perspective and a faster
1.10.2 development cycle.

For the user that is viewing (or possibly reviewing) this commit, I am
going to say that the bulk of the changes here are switches from tabs to
spaces for indentation. If you are wondering, yes, I changed coding
style after I saw that spaces actually looks better on GitHub and offer
better alignment. But that's beside the point.

The newly added @SuppressWarnings are because of the Minecraft
Development plugin, which I admit is really useful, but sometimes a bit
too pedantic. Apart from that, everything's should be golden. The only
TODO that was added is because there is a bit of API missing: I18n
changed the way formatted strings are handled. Now the format strings
are resolved at translation time instead of whenever we wanted like
before. That means a new method should be added to the API to abstract
out this "feature", but that's a change for another time because it's
completely minor: we only use it in one place.

Last thing and then I'm off: porting is now pretty much completed. The
entire architecture of the mod is now ready for use inside a regular
environment. Still, it needs to be completed before we can release 1.0.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>