Porting step 7: Gradle stuff
Updated Gradle to the latest version in order to use multi-threading and
other bits and pieces. Also added license to all files (it was actually
not meant to happen right now, but...).

Checkstyle is also up in place both in IDE and in Gradle so that code
style can be made consistent across the entire code base and PRs.

Findbugs plugin has also been added in a "warning" mode (that is,
errors won't completely destroy the build) to find bugs (duh) in an
easier way.

All the various files affected by checkstyle and license plugin were
also updated (either automatically or manually), so the commit may be a
little bit full of stuff.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>