Porting step 8: Optifine
Write Optifine in-dev integration so that it is easier to test for bugs
that may occur only with it installed (e.g., in 1.7.10 the Sky Block
would glitch out with Shaders, which are now integrated into Optifine).
This allows to mark them note them on the wiki while they wait for a
fix. The Optifine tweaker is enabled by default, but can be disabled
through an in-code flag. It is not exported in the production JAR so
there are no possible conflicts. It needs a custom Optifine version,
refer to simpledeobf for more information.

Also, fixed wrong magic number causing wiki page item list to display
all messed up. Some blocks still like to jump around in the GUI, though.
That is going to get fixed at a later stage.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>