Move full resolution images to dedicated directory
As explained briefly in the relevant README.md in the full_images
directory, moving all the bigger sized images that are mostly not needed
for display in the wiki in a secondary directory gives us the possibility
to exclude them from the final production artifact if the size becomes too
big to handle. Images are always a big problem when they are shaded in
dependencies and this is already a big issue considering the amount of png
files that are present as of now --- considering that the mod is still far
from being actually completed ---. So, better start moving now than later
on.

Also, fixed some documentation issues and explained some decisions and
structures in a better way.

Last but not least, the name of the mod changed from OpenMods-IGW
integration back to the plain OpenMods-IGW, since it is usually referred
to that name as of now, so...

As a note, this could also be referred as OpenWiki or OpenModsWiki, which
in my humble opinion sound a lot better (just writing this for future
reference).

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>